requirements.txt
requirements.txt

numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.0
matplotlib==3.6.2
seaborn==0.12.2
tensorflow==2.11.0
keras==2.11.0
flask==2.2.2
requests==2.28.1
jupyter==1.0.0
pyyaml==6.0
backtrader==1.9.78.123
plotly==5.10.0
pytest==7.2.0
scipy==1.9.3
statsmodels==0.13.5